+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-view/components/add-insight-modal/AddInsightModal.tsx
@@ -1,25 +1,30 @@
-import { FC, useContext, useMemo } from 'react'
+import { FC, ReactElement, ReactNode, useContext, useState, useMemo, FormEvent } from 'react'
 
-import { useQuery } from '@apollo/client'
+import { useApolloClient } from '@apollo/client'
 import { mdiClose } from '@mdi/js'
-import { VisuallyHidden } from '@reach/visually-hidden'
 
-import { isDefined } from '@sourcegraph/common'
-import { Button, Modal, H2, Icon, LoadingSpinner, ErrorAlert } from '@sourcegraph/wildcard'
-
+import { isErrorLike, pluralize } from '@sourcegraph/common'
 import {
-    AccessibleInsight,
-    GetDashboardAccessibleInsightsResult,
-    GetDashboardAccessibleInsightsVariables,
-} from '../../../../../../../graphql-operations'
-import { SubmissionErrors } from '../../../../../components'
+    Button,
+    Modal,
+    H2,
+    Icon,
+    ErrorAlert,
+    MultiCombobox,
+    MultiComboboxInput,
+    MultiComboboxList,
+    MultiComboboxOption,
+    MultiComboboxOptionText,
+    Link,
+} from '@sourcegraph/wildcard'
+
+import { LoaderButton } from '../../../../../../../components/LoaderButton'
+import { GroupByField } from '../../../../../../../graphql-operations'
 import { CodeInsightsBackendContext, CustomInsightDashboard } from '../../../../../core'
+import { encodeDashboardIdQueryParam } from '../../../../../routers.constant'
 
-import {
-    AddInsightFormValues,
-    AddInsightModalContent,
-} from './components/add-insight-modal-content/AddInsightModalContent'
-import { GET_ACCESSIBLE_INSIGHTS_LIST } from './query'
+import { getCachedDashboardInsights, useInsightSuggestions } from './query'
+import { getInsightId, getInsightTitle, InsightSuggestion, InsightType } from './types'
 
 import styles from './AddInsightModal.module.scss'
 
@@ -30,79 +35,186 @@ export interface AddInsightModalProps {
 
 export const AddInsightModal: FC<AddInsightModalProps> = props => {
     const { dashboard, onClose } = props
+
+    const client = useApolloClient()
     const { assignInsightsToDashboard } = useContext(CodeInsightsBackendContext)
 
-    const { data, loading, error } = useQuery<
-        GetDashboardAccessibleInsightsResult,
-        GetDashboardAccessibleInsightsVariables
-    >(GET_ACCESSIBLE_INSIGHTS_LIST, {
-        variables: { id: dashboard.id },
-        errorPolicy: 'all',
-    })
+    const [search, setSearch] = useState('')
+    const [submittingOrError, setSubmittingOrError] = useState<boolean | Error>(false)
+    const [dashboardInsights, setDashboardInsights] = useState(() => getCachedDashboardInsights(client, dashboard.id))
 
-    const insights = getAvailableInsights(data)
-    const dashboardInsightIds = getDashboardInsightIds(data)
+    const excludeIds = useMemo(() => dashboardInsights.map(insight => insight.id), [dashboardInsights])
+    const { connection, loading, fetchMore } = useInsightSuggestions({ search, excludeIds })
 
-    const initialValues = useMemo<AddInsightFormValues>(
-        () => ({
-            searchInput: '',
-            insightIds: dashboardInsightIds,
-        }),
-        [dashboardInsightIds]
-    )
+    const handleSubmit = async (event: FormEvent<HTMLFormElement>): Promise<void> => {
+        event.preventDefault()
+        setSubmittingOrError(true)
 
-    const handleSubmit = async (values: AddInsightFormValues): Promise<void | SubmissionErrors> => {
-        const { insightIds } = values
+        try {
+            const prevInsights = getCachedDashboardInsights(client, dashboard.id)
 
-        await assignInsightsToDashboard({
-            id: dashboard.id,
-            prevInsightIds: dashboardInsightIds,
-            nextInsightIds: insightIds,
-        }).toPromise()
-
-        onClose()
+            await assignInsightsToDashboard({
+                id: dashboard.id,
+                prevInsightIds: prevInsights.map(insight => insight.id),
+                nextInsightIds: dashboardInsights.map(insight => insight.id),
+            }).toPromise()
+            setSubmittingOrError(false)
+            onClose()
+        } catch (error) {
+            setSubmittingOrError(error)
+        }
     }
 
     return (
-        <Modal className={styles.modal} onDismiss={onClose} aria-label="Add insights to dashboard modal">
-            <Button variant="icon" className={styles.closeButton} onClick={onClose}>
-                <VisuallyHidden>Close</VisuallyHidden>
-                <Icon svgPath={mdiClose} inline={false} aria-hidden={true} />
-            </Button>
+        <Modal
+            position="center"
+            aria-label="Add insights to dashboard modal"
+            className={styles.modal}
+            onDismiss={onClose}
+        >
+            <header className={styles.header}>
+                <H2 className="m-0 font-weight-normal">
+                    Add insight to <q>{dashboard.title}</q>
+                </H2>
 
-            {loading && !data && <LoadingSpinner inline={false} />}
-            {error && <ErrorAlert error={error} />}
-            {data && (
-                <>
-                    <H2 className="mb-3">
-                        Add insight to <q>{dashboard.title}</q>
-                    </H2>
+                <Button variant="icon" className={styles.closeButton} aria-label="Close" onClick={onClose}>
+                    <Icon aria-hidden={true} svgPath={mdiClose} />
+                </Button>
+            </header>
 
-                    {!insights.length && <span>There are no insights for this dashboard.</span>}
+            {/* eslint-disable-next-line react/forbid-elements */}
+            <form className={styles.form} onSubmit={handleSubmit}>
+                <MultiCombobox
+                    selectedItems={dashboardInsights}
+                    getItemKey={getInsightId}
+                    getItemName={getInsightTitle}
+                    className={styles.picker}
+                    onSelectedItemsChange={setDashboardInsights}
+                >
+                    <MultiComboboxInput
+                        value={search}
+                        autoFocus={true}
+                        placeholder="Search insights..."
+                        status={loading ? 'loading' : 'initial'}
+                        onChange={event => setSearch(event.target.value)}
+                    />
 
-                    {insights.length > 0 && (
-                        <AddInsightModalContent
-                            initialValues={initialValues}
-                            insights={insights}
-                            dashboardID={dashboard.id}
-                            onCancel={onClose}
-                            onSubmit={handleSubmit}
-                        />
-                    )}
-                </>
-            )}
+                    <small className={styles.description}>
+                        <span>
+                            Don't see an insight?{' '}
+                            <Link to={encodeDashboardIdQueryParam('/insights/create', dashboard.id)}>
+                                Create a new insight.
+                            </Link>
+                        </span>
+                        {connection && (
+                            <span>
+                                {plural('result', connection.nodes.length)}
+                                {connection.totalCount && <> out of {plural('total', connection.totalCount)}</>}
+                            </span>
+                        )}
+                    </small>
+
+                    <MultiComboboxList
+                        items={connection?.nodes ?? []}
+                        renderEmptyList={true}
+                        className={styles.suggestionsList}
+                    >
+                        {items => (
+                            <>
+                                {items.map((item, index) => (
+                                    <InsightSuggestionCard key={getInsightId(item)} item={item} index={index} />
+                                ))}
+                                {items.length === 0 && (
+                                    <span className={styles.zeroStateMessage}>
+                                        {loading ? 'Loading...' : 'No insights found'}
+                                    </span>
+                                )}
+                                {connection?.pageInfo?.hasNextPage && (
+                                    <Button
+                                        size="sm"
+                                        variant="secondary"
+                                        outline={true}
+                                        className={styles.loadMore}
+                                        onClick={fetchMore}
+                                    >
+                                        Load more insights
+                                    </Button>
+                                )}
+                            </>
+                        )}
+                    </MultiComboboxList>
+                </MultiCombobox>
+
+                {isErrorLike(submittingOrError) && <ErrorAlert className="mt-3" error={submittingOrError} />}
+
+                <footer className={styles.footer}>
+                    <span className={styles.keyboardExplanation}>
+                        Press <kbd>↑</kbd>
+                        <kbd>↓</kbd> to navigate through results
+                    </span>
+
+                    <Button
+                        type="button"
+                        className={styles.cancelAction}
+                        variant="secondary"
+                        outline={true}
+                        onClick={onClose}
+                    >
+                        Cancel
+                    </Button>
+
+                    <LoaderButton
+                        alwaysShowLabel={true}
+                        loading={!isErrorLike(submittingOrError) && submittingOrError}
+                        label={!isErrorLike(submittingOrError) && submittingOrError ? 'Saving' : 'Save'}
+                        type="submit"
+                        disabled={!isErrorLike(submittingOrError) && submittingOrError}
+                        variant="primary"
+                    />
+                </footer>
+            </form>
         </Modal>
     )
 }
 
-function getDashboardInsightIds(data?: GetDashboardAccessibleInsightsResult): string[] {
-    if (!data || !data.dashboardInsightsIds.nodes[0]?.views) {
-        return []
-    }
-
-    return data.dashboardInsightsIds.nodes[0].views.nodes.filter(isDefined).map(view => view.id)
+function plural(what: string, count: number): string {
+    return `${count.toLocaleString()} ${pluralize(what, count)}`
 }
 
-function getAvailableInsights(data?: GetDashboardAccessibleInsightsResult): AccessibleInsight[] {
-    return data?.accessibleInsights?.nodes.filter(isDefined) ?? []
+interface InsightSuggestionCardProps {
+    item: InsightSuggestion
+    index: number
+}
+
+function InsightSuggestionCard(props: InsightSuggestionCardProps): ReactElement {
+    const { item, index } = props
+
+    return (
+        <MultiComboboxOption value={getInsightTitle(item)} index={index} className={styles.suggestionCard}>
+            <span>
+                <MultiComboboxOptionText />
+            </span>
+            <small className={styles.suggestionCardDescription}>
+                {item.type} insight {getInsightDetails(item)}
+            </small>
+        </MultiComboboxOption>
+    )
+}
+
+function getInsightDetails(insight: InsightSuggestion): ReactNode {
+    switch (insight.type) {
+        case InsightType.Detect:
+            return insight.queries.join(', ')
+        case InsightType.DetectAndTrack:
+            return insight.query
+        case InsightType.Compute:
+            return `${insight.query}, grouped by ${formatGroupBy(insight.groupBy)}`
+        case InsightType.LanguageStats:
+            return ''
+    }
+}
+
+const formatGroupBy = (groupBy: GroupByField): string => {
+    const str = groupBy.toString()
+    return str.charAt(0).toUpperCase() + str.slice(1).toLowerCase()
 }